Move CrossRef from simple license info to ListedLicense
